Save whether "show pictures" has been clicked
Save/restore whether "show pictures" has been clicked. On restore this
allows us to correctly instruct MessageContainerView whether to load
pictures.
